Moved gzip to its own header and renamed the remainder of utility to conversions.
